requirements.txt
python-dotenv
nova-act
boto3
requests
